requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
